+++ app/src/main/java/mb28/crysongs/ui/popups/TrackMoreOptionsPopup.kt
@@ -1,5 +1,6 @@
 package mb28.crysongs.ui.popups
 
+import android.widget.Toast
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.OutlinedButton
@@ -9,12 +10,14 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.platform.LocalContext
 import mb28.crysongs.core.Settings
 import mb28.crysongs.core.Track
 import mb28.crysongs.icons.favorite
 import mb28.crysongs.icons.heart_plus
 import mb28.crysongs.icons.playlist_add
 import mb28.crysongs.icons.queue_music
+import mb28.crysongs.nowPlaying
 import mb28.crysongs.nowPlayingI
 import mb28.crysongs.playerQuery
 import mb28.crysongs.setAndPlay
@@ -23,6 +26,7 @@ import mb28.crysongs.updateDisplayQuery
 
 @Composable
 fun TrackMoreOptionsPopup(track: Track, onDismissRequired: () -> Unit) {
+    val context = LocalContext.current
     var showAddToPL by remember { mutableStateOf(false) }
     AlertDialog(
         { onDismissRequired() },
@@ -54,11 +58,17 @@ fun TrackMoreOptionsPopup(track: Track, onDismissRequired: () -> Unit) {
                         "Play next",
                         2, count,
                     ) {
-                        if (playerQuery.isNotEmpty()) {
-                            playerQuery.add(nowPlayingI + 1, track)
-                        } else {
+                        if (playerQuery.isEmpty()) {
                             playerQuery.add(track)
                             setAndPlay(track, false)
+                        } else if (nowPlaying == track) {
+                            Toast.makeText(context, "Already playing!", Toast.LENGTH_SHORT).show()
+                        } else if (playerQuery.contains(track)) {
+                            playerQuery.run {
+                                add(nowPlayingI + 1, removeAt(indexOf(track)))
+                            }
+                        } else {
+                            playerQuery.add(nowPlayingI + 1, track)
                         }
                         updateDisplayQuery()
                         onDismissRequired()